# Compatibility with external deep proxies