… package.json
… test/spec/proxySpec.js