@tomalec 2019-06-06
Could you add a test for creating external deep proxy?
I still don't see a clear reason why previous implementation with a map is not capable of doing so. What's the difference between of having instance specific map than instance specific symbol?