@warpech 2019-06-06
> Could you add a test for creating external deep proxy?

There is a test in the commit c461b492a4fa2e2869ba99cbe0329db3846e93cd

> I still don't see a clear reason why previous implementation with a map is not capable of doing so. What's the difference between of having instance specific map than instance specific symbol?

Consider this. There are two ways of accessing an object:
- by reference. In JSONPatcherProxy, such way of accessing is used whenever you see `tree`. This way gives you original unproxified instance of the object, because that's how the `set` trap works.
- by parent and key. In JSONPatcherProxy, such way of accessing is used whenever you see `tree[key]`. This way gives you whatever the external world sees as the value, including the proxy made by the current instance of JSONPatcherProxy and any other proxy.

All places in JSONPatcherProxy where there was a need to use `tree[key]`, used to use `tree[key]` as the key to a `Map` object. However, this only worked as long as the value at `tree[key]` was exactly the proxy object created by the current instance of JSONPatcherProxy. It the value was additionaly proxified by external code, it was detected by JSONPatcherProxy as a totally new object.

This is because in the ES Proxy world, the proxy instances have different identity than the original object:

```
var obj = {};
var p1 = new Proxy(obj, {});
var p2 = new Proxy(p1, {});
obj === p1; //false
p1 === p2; //false
```

The method used in this PR works by recognizing the object already proxified by an instance of JSONPatcherProxy not by keeping a reference to it, but by marking it from the inside:

```
var obj = {};
var meta = {};
var symb = Symbol("Symbol for getting the tree metadata from external access.");
obj[symb] = meta;

var p1 = new Proxy(obj, {});
var p2 = new Proxy(p1, {});
obj[symb] === p1[symb]; //true
p1[symb] === p2[symb]; //true
```