@tomalec 2019-06-06
> There is a test in the commit c461b49

Sorry my bad

----

Thanks, now I get why having Map of object references will not work.
But then doesn't keeping `parent` object reference shares the same problem? If it doens't maybe we could keep a map of `parent, key` tuples.

I'm a little bit picky, as I remember we used to have implementation that was using `get` trap to identify proxified object, and we discarded it for some reason (I suspect performance)