@tomalec 2019-06-07
Have you tried 
> But then doesn't keeping parent object reference shares the same problem? If it doesn't maybe we could keep a map of parent, key tuples.