@warpech 2019-06-07
I pushed an additional test to check if double proxification by `JSONPatcherProxy` works as intended. It does work.