@warpech 2019-06-24
I made a new PR that contains only the test refactorings from the current PR: https://github.com/Palindrom/JSONPatcherProxy/pull/42