@warpech 2019-06-24
This PR was rebased on current `master` and is now available as a new PR #43